Rename Netty utils to cookie encoding (#7887)
Fixes #7554.

The Netty utils classes were copied from Netty but since they only contain cookie encoding logic it was confusing to have the name Netty in the project dependencies.
