+++ framework/src/play/src/main/java/play/core/cookie/encoding/CookieDecoder.java
@@ -13,16 +13,14 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package play.core.netty.utils;
+package play.core.cookie.encoding;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static play.core.netty.utils.CookieUtil.firstInvalidCookieNameOctet;
-import static play.core.netty.utils.CookieUtil.firstInvalidCookieValueOctet;
-import static play.core.netty.utils.CookieUtil.unwrapValue;
-
 import java.nio.CharBuffer;
+
+import static play.core.cookie.encoding.CookieUtil.*;
 
 /**
  * Parent of Client and Server side cookie decoders
