+++ framework/src/play/src/main/java/play/core/cookie/encoding/ServerCookieEncoder.java
@@ -13,15 +13,11 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package play.core.netty.utils;
+package play.core.cookie.encoding;
 
-import static play.core.netty.utils.CookieUtil.*;
+import java.util.*;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
+import static play.core.cookie.encoding.CookieUtil.*;
 
 /**
  * A <a href="https://tools.ietf.org/html/rfc6265">RFC6265</a> compliant cookie encoder to be used server side,
